+++ pkgs/ffigen/lib/src/config_provider/config.dart
@@ -176,6 +176,14 @@ class Config {
   FfiNativeConfig get ffiNativeConfig => _ffiNativeConfig;
   late FfiNativeConfig _ffiNativeConfig;
 
+  /// Where to ignore compiler warnings/errors in source header files.
+  bool get ignoreSourceErrors => _ignoreSourceErrors;
+  set ignoreSourceErrors(bool n) {
+    _ignoreSourceErrors = _ignoreSourceErrors || n;
+  }
+
+  bool _ignoreSourceErrors = false;
+
   Config._({required this.filename, required this.packageConfig});
 
   /// Create config from Yaml map.
@@ -285,6 +293,12 @@ class Config {
               transform: (node) => headersExtractor(node.value, filename),
               result: (node) => _headers = node.value,
             )),
+        HeterogeneousMapEntry(
+          key: strings.ignoreSourceErrors,
+          valueConfigSpec: BoolConfigSpec(),
+          defaultValue: (node) => false,
+          resultOrDefault: (node) => ignoreSourceErrors = node.value as bool,
+        ),
         HeterogeneousMapEntry(
           key: strings.compilerOpts,
           valueConfigSpec: OneOfConfigSpec<List<String>, List<String>>(