+++ pkgs/ffigen/lib/src/header_parser/utils.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:ffi';
+import 'dart:math' show max;
 
 import 'package:ffi/ffi.dart';
 import 'package:ffigen/src/code_generator.dart';
@@ -29,16 +30,21 @@ void visitChildrenResultChecker(int resultCode) {
 }
 
 /// Logs the warnings/errors returned by clang for a translation unit.
-void logTuDiagnostics(
+///
+/// Returns the highest [clang_types.CXDiagnosticSeverity] seen, defaults to
+/// [clang_types.CXDiagnosticSeverity.CXDiagnostic_Ignored] if none.
+int logTuDiagnostics(
     Pointer<clang_types.CXTranslationUnitImpl> tu, Logger logger, String header,
     {Level logLevel = Level.SEVERE}) {
+  var result = clang_types.CXDiagnosticSeverity.CXDiagnostic_Ignored;
   final total = clang.clang_getNumDiagnostics(tu);
   if (total == 0) {
-    return;
+    return result;
   }
   logger.log(logLevel, 'Header $header: Total errors/warnings: $total.');
   for (var i = 0; i < total; i++) {
     final diag = clang.clang_getDiagnostic(tu, i);
+    result = max(result, clang.clang_getDiagnosticSeverity(diag));
     final cxstring = clang.clang_formatDiagnostic(
       diag,
       clang_types
@@ -50,6 +56,7 @@ void logTuDiagnostics(
     logger.log(logLevel, '    ${cxstring.toStringAndDispose()}');
     clang.clang_disposeDiagnostic(diag);
   }
+  return result;
 }
 
 extension CXSourceRangeExt on Pointer<clang_types.CXSourceRange> {